Move merge fns.
